requirements.txt
beautifulsoup4
pillow
python-telegram-bot
requests
rich
typer
